[#114] :bug: fix: PlanService < 통합
